+++ lib/runners/java.rb
@@ -2,6 +2,13 @@ module Runners
   module Java
     class InvalidDependency < UserError; end
 
+    def self.included(klass)
+      # @type var klass: singleton(Processor)
+      Config.register_warnings do |config|
+        config.add_warning_for_deprecated_option(analyzer: klass.analyzer_id, old: :jvm_deps, new: :dependencies)
+      end
+    end
+
     def show_runtime_versions
       capture3! "java", "-version"
       capture3! "gradle", "--version"
@@ -15,7 +22,7 @@ module Runners
 
         trace_writer.message "Install dependencies..." do
           fetch_deps_via_gradle!
-          deps = config_jvm_deps.map { |dep| "* #{dep.join(':')}" }
+          deps = config_jvm_deps.map { |dep| "* #{dep}" }
           trace_writer.message <<~MSG
             Successfully installed #{deps.size} #{deps.size == 1 ? 'dependency' : 'dependencies'}:
             #{deps.join("\n")}
@@ -31,7 +38,21 @@ module Runners
     end
 
     def config_jvm_deps
-      @config_jvm_deps ||= Array(config_linter[:jvm_deps]).each do |dep|
+      return config_jvm_deps_old unless config_linter[:dependencies]
+
+      @config_jvm_deps ||= Array(config_linter[:dependencies]).map do |dep|
+        # @type var dep: Hash[Symbol, String] | String
+        case dep
+        when Hash
+          dep.values.join(":")
+        else
+          dep
+        end
+      end
+    end
+
+    def config_jvm_deps_old
+      @config_jvm_deps_old ||= Array(config_linter[:jvm_deps]).map do |dep|
         # @type var dep: Array[String]
         group, name, version = dep
         unless group && name && version
@@ -42,6 +63,8 @@ module Runners
           trace_writer.error message
           raise InvalidDependency, message
         end
+
+        dep.join(":")
       end
     end
 
@@ -52,7 +75,7 @@ module Runners
       repos = ["mavenCentral()", "jcenter()", "google()"].map { |repo| "  #{repo}" }.join("\n")
 
       trace_writer.message "Generating #{filename}..." do
-        deps = config_jvm_deps.map { |dep| "  implementation '#{dep.join(":")}'" }.join("\n")
+        deps = config_jvm_deps.map { |dep| "  implementation '#{dep}'" }.join("\n")
         content = <<~GRADLE
           plugins {
             id 'java'