+++ test/smokes/checkstyle/dependencies/Foo.java
@@ -1,0 +1,3 @@
+@XXX
+class Foo {
+}